+++ src/app/progress/milestone-items.ts
@@ -275,6 +275,7 @@ function makeFakePursuitItem(
     ammoType: DestinyAmmunitionType.None,
     missingSockets: false,
     breakerType: null,
+    foundry: null,
     pursuit: null,
     taggable: false,
     comparable: false,